Add state_class to Battery Level Sensor
I like to add long term statistics to the Battery Level Sensor in order to recognize battery health.

Home Assistant has support for storing sensors as long-term statistics if the entity has the right properties. To opt-in for statistics, the sensor must have state_class set to one of the valid state classes: SensorStateClass.MEASUREMENT, SensorStateClass.TOTAL or SensorStateClass.TOTAL_INCREASING. For certain device classes, the unit of the statistics is normalized to for example make it possible to plot several sensors in a single graph. 

(https://developers.home-assistant.io/docs/core/entity/sensor/#long-term-statistics)